requirements.txt
python-telegram-bot[fast]>=22.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
supabase>=2.7.0
python-dotenv>=1.0.1
httpx>=0.27.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
tenacity>=8.2.0
twilio>=9.0.0
